Many changes done to get updating of travel times to fully work. Changed how db sessions are handled so that data is successfully written to database. This is critical because previously was calling saveOrUpdate() on data that was read in, which was causing problems where Hibernate was deleting stop paths from trip patterns for no apparent reason. Instead, just need to make necessary changes within a transaction and when the transaction and session are closed then all "dirty" objects are automatically written.
Also greatly changed how the active revision is handled. No more sandbox
rev. Instead, a new revision is written with a new configRev and it can
be switched to via ActiveRevisions db table without having to copy the
revision.  This greatly simplifies handling of revisions.

Also, fixed some issues with updating of travel times where had some
bugs that were causing the data to not all be handled correctly.