Changed significantly how WebAgency objects are read in and cached. Now can better control better when WebAgencies are read from db again. When an RMI method is invoked and there is a problem then WebAgencies read in again almost right away (as long as least 30 seconds old) so that if server hostname changed it will be picked up and will automatically connect to the new host. So can more easily move agencies from one server to another.
Also changed how an ordered list of WebAgencies is provided. Now the
agency name is cached so don't query the name via RMI, as was done
before. This eliminates problem when an agency is not working and
getting the name would result in a 4-8 second timeout and if frequently
getting the name to do a sort it could take forever to finish.

And changed so that when simply creating a client factory it uses cached
WebAgencyc data since it is most likely correct. This speeds up these
first calls.