Improved AutoBlockAssigner feature so that it runs a magnitude faster. Previously was taking 5-10 seconds when auto assigning and there were a large number of blocks available to be assigned. This was noticed for VTA since it is a large agency with many blocks, the AVL feed doesn't have assignment info necessitating the auto assigner, the AVL feed reports very frequently (every second!), and only a few buses currently have GPS meaning that there will always be a large number of available blocks.
The auto assignment feature was taking so long that processing AVL data
was getting backed up when running on a laptop. Simply not enough
processing power to handle an inefficient algorithm. Since AVL was
getting further and further behind pretty quickly the system stopped
working.

Changed to cache searches for matches by trip pattern. This way only
match to a trip pattern once. Also added a great deal of debug logging
to confirm that system was working. This reduced auto assigning a
vehicle to about 400msec. Also, changed so that don't try to autoassign
to a trip if AVL report not even within the extent of the trip. This
greatly reduces number of trips/trip patterns that need to be
investigated.

This improvement required changes in SpatialMatch class so affected
multiple classes.