Changed temporal matching to handle vehicles leaving early from a wait stop. There is a complication with vehicles leaving a layover slightly early where might not temporally match to a match past the layover due to system thinking that vehicle first needs to go back to the layover and then to the spatial match. This leads to an overly long expected travel time and the result would be that the system thinks the vehicle is at the match earlier than expected. At the same time it will think that it is late with respect to the layover, but since late is considered a better match than early vehicle could be wrongly temporally matched to layover again. As the vehicle moves further along the trip this only gets worse since system will expect a longer and longer travel time for vehicle to go back to layover and then to the spatial match. So could get incorrect layover match for quiet a while, until the layover distance transitime.core.layoverDistance, which can be large, is exceeded.
Therefore need to filter out such problematic layover matches. This is
done by seeing if past schedule time for the layover match (before the
schedule time vehicles should be able to roam around), the previous
match for vehicle is same layover as the current spatial match, the
distance from the layover stop is increasing, and there is a subsequent
non-layover spatial match that could be used. If all these conditions
are true then the current layover spatial match is ignored so that will
properly match vehicle to subsequent non-layover spatial match.

Also changed genernation of predictions for a wait stop such that a
configuration parameter useExactSchedTimeForWaitStops tells system to
use schedule time instead of actual predicted time for predictions
provided to users. This way even if historic data indicates vehicles
leave a bit later than schedule the system will still provide the
schedule time to the user since that is more conservative and is
probably what is expected. This avoids problem of doors closing on
schedule but departure not being recorded until a bit later. The actual
prediction time based on historic departure times is still used for
generating the predictions for subsequent stops so that they are as
accurate as possible.