Made many changes to 1) improve handling of route names; and 2) make API more consistent.
With respect to route name changes, now keep track of the GTFS long
namem as well so it can be accessed, including via the API. This way a
UI can get access to the names before they have been processed. Also
Changed creation of the route name so that it is always automatically
determined (don't need to use the -combineRouteNames anymore). This way
the full name is always available. The database table for Routes was
also modified to accomodate this change.

With respect to API, changed getting of routes to have the commands
"routes and "routesDetails" so consistent with "vehicles" and
"vehiclesDetails". Changed getting detailed route info so that always
get an array back, even if just getting data for a single route. This
makes the call more consistent.