Added socket timeout
For when session lost due to db restart. 

NOTE: this change was created in the main Transitime branch for the DataDbLogger class. But simultaneously CS changed db logging to use this new DbQueue class. Therefore had to try to merge the changes that were tested in the main branch with this DbQueue class. This means that there might be issues with this code since I am not familiar with DbQueue class.