Updates to PlaybackModule.
Usage:

Run the class org.transitime.playback.PlaybackModule with arguments <path-to-gtfs-folder> <path-to-csv>, where gtfs-folder is a folder
of unzipped GTFS files (probably restricted to a particular route) and the csv is an export from the
AVL reports table. Use a SQL statement like the following:

 select vehicleId, time, assignmentId, assignmentType, heading, lat as latitude, lon as longitude
  from AvlReports
  where time between "2016-04-21 18:30:00" and "2016-04-21 20:00:00"
  and vehicleId="8087"

There is example GTFS folder (WMATA 5A) and AVL CSV trace in src/main/resources.

I run the class as an Eclipse run configuration. I typically examine a trace by looking at the logs,
inspecting the database, and/or stepping through AvlProcessor.

Logs by default go to /Logs/DEFAULT/core/YYYY/MM/DD.

The database is in-memory HSQL by default. If you want to inspect the database, it has to be an out-of-process
HSQL database:

1) In src/main/resources/hsql_integration_test_hibernate.cfg.xml, change hibernate.connection.url to
jdbc:hsqldb:hsql://localhost/xdb

2) Download hsqldb.jar from http://hsqldb.org/download/

3) Start HSQLDB: java -cp hsqldb.jar org.hsqldb.server.Server --database.0 file:mydb --dbname.0 xdb

4) Inspect the database with the database manager:
 java -cp hsqldb.jar org.hsqldb.util.DatabaseManagerSwing

TODO (incomplete):
- Make it configurable which kind of HSQL database is used.
- Configurable what steps of playback run.
- Delete stateful database tables before run so we can use an old database.
- Make into an integration test.